correct project structure
