+++ src/main/java/net/aether/lib/data/DoubleHashMap.java
