+++ src/main/java/net/aether/lib/data/SimpleQueue.java
