+++ src/main/java/net/aether/lib/data/Void.java
