+++ src/main/java/net/aether/lib/exceptions/InvalidTypeException.java
