+++ src/main/java/net/aether/lib/io/FileContentReader.java
