+++ src/main/java/net/aether/lib/lambda/Consumer.java
