+++ src/main/java/net/aether/lib/lambda/Executable.java
