+++ src/main/java/net/aether/lib/lambda/TriConsumer.java
