+++ src/main/java/net/aether/lib/lambda/TriConverter.java
