+++ src/main/java/net/aether/lib/misc/ArrayUtils.java
