+++ src/main/java/net/aether/lib/misc/StaticLoggerUtils.java
