+++ src/main/java/net/aether/lib/misc/StringUtils.java
